LIBITD-938 Add ability to copy records from archive
This allows users to go to archived records and "copy" them to new
records in the current FY.

https://issues.umd.edu/browse/LIBITD-938
