+++ app/assets/javascripts/application.js
@@ -25,3 +25,4 @@
 //= require toggle    
 //= require unit_selector 
 //= require word_limit 
+//= require spawner
