+++ app/views/home/index.html.erb
@@ -6,6 +6,7 @@
     <li><%= link_to( t('actions.new') + " Contractor Request", new_polymorphic_url( ContractorRequest ), class: "btn btn-warning" ) %></li> 
     <li>Version: <%= APP_VERSION %></li>
     <li>Tag: <%= GIT_TAG %></li>
+    <li>Current Fiscal Year: <%= FISCAL_YEAR %></li>
   </ul>
 </div>
