+++ app/views/requests/_labor_request_form.html.erb
@@ -1,25 +1,26 @@
-<%= simple_form_for @request do |f| %>
+<%= simple_form_for @request, html:  {"data-model-klass": @model_klass.name.underscore }   do |f| %>
+
   
   <%= render partial: "edit_form_header", locals: { f: f } %> 
 	<%= render 'shared/form_errors', errors: @request.errors %>
 
   <div class='panel panel-default'>
     <div class="panel-body">
-      <%= f.input :position_title %>
-      <%= f.input :employee_type, as: :select,  collection:  @model_klass::VALID_EMPLOYEE_TYPES %>
-      <%= f.input :request_type, as: :select,  collection:  @model_klass::VALID_REQUEST_TYPES %>
-      <%= f.input :contractor_name %>
-      <%= f.input :number_of_positions %>
-      <%= f.input :hourly_rate, value: number_with_precision( f.object.hourly_rate, precision: 2, delmitier: ',' ) %>
-      <%= f.input :hours_per_week %>
-      <%= f.input :number_of_weeks %>
-      <%= f.input :nonop_funds %>
-      <%= f.input :nonop_source %>
+      <%= f.input :position_title, input_html: { class: 'copy' } %>
+      <%= f.input :employee_type, as: :select,  collection:  @model_klass::VALID_EMPLOYEE_TYPES , input_html: { class: 'copy' } %>
+      <%= f.input :request_type, as: :select,  collection:  @model_klass::VALID_REQUEST_TYPES , input_html: { class: 'copy' } %>
+      <%= f.input :contractor_name , input_html: { class: 'copy' } %>
+      <%= f.input :number_of_positions , input_html: { class: 'copy' } %>
+      <%= f.input :hourly_rate, value: number_with_precision( f.object.hourly_rate, precision: 2, delmitier: ',' ) , input_html: { class: 'copy' } %>
+      <%= f.input :hours_per_week , input_html: { class: 'copy' } %>
+      <%= f.input :number_of_weeks , input_html: { class: 'copy' } %>
+      <%= f.input :nonop_funds , input_html: { class: 'copy' } %>
+      <%= f.input :nonop_source , input_html: { class: 'copy' } %>
       
-      <%= f.association :organization, collection: current_user.available_departments, label_method: :name, input_html: { class: 'department-select' }, include_blank: false %>
-      <%= f.association :unit, collection: current_user.active_units, input_html: { class: 'unit-select' }  %>
+      <%= f.association :organization, collection: current_user.available_departments, label_method: :name, input_html: { class: 'department-select copy' }, include_blank: false %>
+      <%= f.association :unit, collection: current_user.active_units, input_html: { class: 'unit-select copy' }  %>
 
-      <%= f.input :justification, input_html: { class: "word-limit", "data-word-limit": 125 } %>
+      <%= f.input :justification, input_html: { class: "word-limit copy", "data-word-limit": 125 } %>
       
     </div>
   </div>
