+++ config/initializers/version.rb
@@ -1,2 +1,3 @@
 APP_VERSION = '1.2.0-19-g29d4d12'
 GIT_TAG = '29d4d12e6a7963734b3cf27003c8807977acdedd'
+FISCAL_YEAR= 'FY19'
